safari-600.7-branch/LayoutTests/svg/animations/animVal-basics-expected.txt

SVG 1.1 dynamic animation tests

Trivial animVal testcase, to see wheter we support it at all. Should result in a 200x200 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 200
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 150
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 200
PASS rect.width.animVal.value is 200
PASS rect.width.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
